add Tracer trace_ray function with tmin parameter

+++ Src/Tracer.cpp
@@ -44,7 +44,7 @@ Tracer::trace_ray(const Ray ray, const int depth) const {
 	return (black);
 }
 
-RGBColor Tracer::trace_ray(const Ray ray, float& tmin, const int depth) const{
+RGBColor Tracer::trace_ray(const Ray ray, double& tmin, const int depth) const{
 	return (black);
 }
