+++ Src/Tracer.h
@@ -34,7 +34,7 @@ class Tracer {
 
 		virtual RGBColor trace_ray(const Ray ray, const int depth) const;
 
-		virtual RGBColor trace_ray(const Ray ray, float& tmin, const int depth) const;
+		virtual RGBColor trace_ray(const Ray ray, double& tmin, const int depth) const;
 				
 	protected:
